Recycle unused pointers in a "freelist"
There are a lot of cases where we end up creating and deleting a bunch
of extraneous objects. While optimizing that would be nice, in the mean
time, we can just save them to a large array of available entries, and
then try to reuse them for later allocations of the same size. This can
result in some absolutely massive speed-ups.